requirements.txt
matplotlib==3.10.8
matplotlib-inline==0.2.1
numpy==1.26.4
pillow==12.0.0
torch==2.7.0+cu126
torchaudio==2.7.0+cu126
torchmetrics==1.8.2
torchvision==0.22.0+cu126
